feat(session): swap old session entity for new view-based design
Replace the old entity/thread system (which duplicated full prompt data
on threads) with the new view-based design where all data lives as
session-level events and threads hold lightweight index-based views.

- Rename new_entity.rs → entity.rs, new_thread.rs → thread.rs
- Add From conversions between message and llm types at service boundary
- Rewrite Sessions service with new API (add_user_input, next_prompt,
  assistant_response_received, add_tool_results)
- Rewrite agent loop to use response-driven flow
  (PromptPending/ToolUseRequest/Done)
- Restore accidentally removed tool_choice field on llm::Prompt
