+++ core/tests/prompt_executor.rs
@@ -33,6 +33,7 @@ async fn anthropic_round_trip_via_executor() {
         }],
         system: Vec::new(),
         tools: Vec::new(),
+        tool_choice: None,
         max_tokens: None, // executor should fill this in from default_max_tokens
     };
